refactor(substrate): extract shared tick_only_lifecycle_config helper
Qodana flagged the identical Tick-self-loop graph + relay-subscriber
block duplicated across the headless, test_bench, and desktop chassis
builders. Pull it into `chassis_common::tick_only_lifecycle_config()`
so the three chassis call one helper. A chassis that later adds
Render/Present stages (ADR-0082 §11) builds its own graph instead.
